requirements.txt
fastapi
uvicorn[standard]
python-dotenv
llama-index>=0.10.0 # Use a recent version
llama-index-llms-openai
llama-index-embeddings-openai
llama-index-vector-stores-chroma
llama-index-readers-file # Includes PDF and Excel readers
chromadb>=0.4.22 # Use a compatible version
pypdf # Dependency for PDF reader
openpyxl # Dependency for XLSX reader
pandas # Dependency for XLSX reader